+++ android/src/main/kotlin/com/bitmovin/player/flutter/json/JsonMethodHandler.kt
@@ -4,7 +4,7 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 
 /** set Result from returned object or set an error if an exception is thrown. */
-internal class JsonMethodHandler(private val handler: (String, JMethodArgs) -> Any) :
+internal class JsonMethodHandler(private val handler: (String, JMethodArgs) -> Any?) :
     MethodChannel.MethodCallHandler {
     override fun onMethodCall(
         call: MethodCall,